Fix scrub import again (#30)
* reversed imports

* bumped version

* added scrub to exports
