+++ templates/axios.js
@@ -5,7 +5,6 @@ const https = require('https');
 const dotenv = require('dotenv');
 const polyCustom = require('./poly-custom');
 const { API_KEY, API_BASE_URL } = require('./constants');
-const { scrub } = require('../api/index.js')
 
 dotenv.config();
 
@@ -43,6 +42,29 @@ axios.interceptors.request.use(
   }
 );
 
+
+const scrub = (data) => {
+  if (!data || typeof data !== 'object' ) return data;
+  const secrets = ["x_api_key", "x-api-key", "access_token", "access-token", "authorization", "api_key", "api-key", "apikey", "accesstoken", "token", "password", "key"];
+  if (Array.isArray(data)) {
+    return data.map(item => scrub(item))
+  }
+  else {
+    const temp = {};
+    for (const key of Object.keys(data)) {
+      if (typeof data[key] === 'object') {
+        temp[key] = scrub(data[key]);
+      } else if (secrets.includes(key.toLowerCase())) {
+        temp[key] = "********";
+      } else {
+        temp[key] = data[key];
+      }
+    }
+    return temp
+  }
+}
+
+
 const scrubKeys = (err) => {
   if (!err.request || typeof err.request.headers !== 'object') throw err
   const temp = scrub(err.request.headers)
@@ -59,5 +81,6 @@ const scrubKeys = (err) => {
 
 module.exports = {
   axios,
-  scrubKeys
+  scrubKeys,
+  scrub
 };
